introduced defaultVoc.js
